+++ include/symbol_table.h
@@ -10,28 +10,28 @@ enum class SymbolType {
 };
 
 struct SymbolEntry {
-    double value;
+    long double value;
     SymbolType type;
 
     SymbolEntry() : value(0.0), type(SymbolType::VARIABLE) {}  // Default constructor with initial values
-    SymbolEntry(double val, SymbolType t) : value(val), type(t) {}
+    SymbolEntry(long double val, SymbolType t) : value(val), type(t) {}
 };
 
 
 class SymbolTable {
 public:
     // Methods to declare constants and variables
-    void declareConstant(const std::string& name, double value);
-    void declareVariable(const std::string& name, double value, double skipCheck = false);
+    void declareConstant(const std::string& name, long double value);
+    void declareVariable(const std::string& name, long double value, long double skipCheck = false);
     // Lookup for a symbol (constant or variable)
-    double lookupSymbol(const std::string& name) const;
+    long double lookupSymbol(const std::string& name) const;
 
     void clearVariables();
-    void restoreVariables(const std::unordered_map<std::string, double>& savedVariables);
+    void restoreVariables(const std::unordered_map<std::string, long double>& savedVariables);
 
-    std::unordered_map<std::string,double> getConstants() const;
+    std::unordered_map<std::string,long double> getConstants() const;
 
-    std::unordered_map<std::string,double> getVariables() const;
+    std::unordered_map<std::string,long double> getVariables() const;
 
     std::unordered_map<std::string, SymbolEntry> getEntries() const;
